2019: Day 05/07/09: IntComputer handle longs

+++ 2019/src/main/java/com/augustoccesar/adventofcode/day07/Circuit.java
@@ -24,7 +24,7 @@ public class Circuit {
         });
   }
 
-  public int run() {
+  public long run() {
     while (true) {
       IntComputer currentAmplifier = this.amplifiers.get(this.currentAmplifier);
       currentAmplifier.run();
